Create formatting-tree directory
