+++ ts/domlib/surround/evaluate-format.ts
@@ -1,8 +1,8 @@
 // Copyright: Ankitects Pty Ltd and contributors
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 
+import type { FormattingNode } from "./formatting-tree";
 import type { SurroundFormat } from "./match-type";
-import type { FormattingNode } from "./tree-node";
 
 export class EvaluateFormat {
     constructor(protected format: SurroundFormat) {}
