+++ ts/domlib/surround/formatting-tree/index.ts
@@ -1,0 +1,7 @@
+// Copyright: Ankitects Pty Ltd and contributors
+// License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
+
+export { BlockNode } from "./block-node";
+export { FormattingNode } from "./formatting-node";
+export { MatchNode } from "./match-node";
+export type { TreeNode } from "./tree-node";
