+++ ts/domlib/surround/match-type.ts
@@ -1,6 +1,9 @@
 // Copyright: Ankitects Pty Ltd and contributors
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 
+import { nodeIsCommonElement } from "../../lib/dom";
+import type { FormattingNode, MatchNode } from "./formatting-tree";
+
 export enum MatchType {
     /**
      * An element unrelated to the surround format.
@@ -62,8 +65,6 @@ export interface FoundMatch {
 
 // export type RangeMerger = (before: ChildNodeRange, after: ChildNodeRange) => boolean;
 
-import { nodeIsCommonElement } from "../../lib/dom";
-import type { FormattingNode, MatchNode } from "./tree-node";
 // TODO REMOVE
 /**
  * We want to avoid that users have to deal with the difference between Element
